added pull impl

+++ include/uvw/variable.h
@@ -18,40 +18,82 @@ namespace uvw
     friend class Workspace;
     friend class Processor;
 
+    protected:
+
     Duo key_;
     Duo src_;
     size_t type_; // std::typeinfo::hash_code()
-
-    protected:
+    void* data_ptr_;
+    void* data_src_;
 
     Variable(
       const Duo& key,
       size_t type_code = 0
-    ): key_(key), type_(type_code) {}
+    ): key_(key), type_(type_code)
+    {
+      data_ptr_ = data_src_ = nullptr;
+    }
 
     public:
-    Variable(): type_(0) {}
+    Variable(): type_(0)
+    {
+      data_ptr_ = data_src_ = nullptr;
+    }
 
     const Duo& key() {return key_;}
-    const std::string label() {return key_.var;}
+    const std::string label() {return key_.var_str;}
     Processor* proc();
 
     template<typename T> bool is_of_type();
     template<typename T> static bool is_null(const T& obj);
 
-    void unlink() {src_ = Duo(nullptr,"");}
-    bool link(const Variable& src);
-    const Duo source() {return src_;}
+    void unlink() {src_ = Duo(nullptr,""); data_src_ = nullptr;}
+    bool link(Variable* src);
+    const Duo src() {return src_;}
+
+    virtual void pull() {}
 
     protected:
     template<class T> static T null_;
     template<class T> static size_t var_type;
   };
 
+  template<class T>
+  class Var : public Variable
+  {
+    friend class Workspace;
+    friend class Processor;
+
+    protected:
+
+    T value_;
+
+    Var(const Duo& key):
+      Variable(key, typeid(T).hash_code()) {}
+
+    public:
+
+    Var(): Variable() {}
+
+    void pull() override
+    {
+      if (data_src_)
+      {
+        value_ = *((T*)data_src_);
+      }
+    }
+
+    // type-specific members
+    T& ref() {return value_;}
+    T& operator()() {return value_;}
+    void set(const T& val) {value_ = val;}
+    T get() {return value_;}
+  };
+
   // impl.
 
   template<class T> T Variable::null_;
-  template<class T> size_t Variable::var_type = typeid(Variable::null_<T>).hash_code();
+  template<class T> size_t Variable::var_type = typeid(T).hash_code();
   template<typename T> bool Variable::is_null(const T& obj)
   {
     return &obj == &Variable::null_<T>;
@@ -61,14 +103,16 @@ namespace uvw
     return (var_type<T> == type_);
   }
 
-  bool Variable::link(const Variable& src)
+  bool Variable::link(Variable* src)
   {
-    if (type_ != src.type_)
+    if (type_ !=0 && type_ != src->type_)
     {
       unlink();
       return false;
     }
-    src_ = Duo(src.key_);
+    // hook up key & src data ptr
+    src_ = Duo(src->key());
+    data_src_ = src->data_ptr_;
     return true;
   }
